requirements.txt
-i https://pypi.org/simple
certifi==2018.8.24
chardet==3.0.4
idna==2.7
kafka-python==1.4.3
requests==2.19.1
urllib3==1.23; python_version != '3.1.*'
